refactor: move module `url` into `helper`
